requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
Django==4.1.7
django-bootstrap-v5==1.0.11
django-filter==22.1
django-tables2==2.5.3
mysqlclient==2.1.1
soupsieve==2.4
sqlparse==0.4.3
tzdata==2023.2
